chunk,serializable update

+++ source/include/tileMap.hpp
@@ -6,7 +6,7 @@ class TileMap : public sf::Drawable, public sf::Transformable
 public:
 
     bool load(const std::string& tileset, sf::Vector2u tileSize,
-			  int* tiles, unsigned int width, unsigned int height);
+			  uint16_t* tiles, unsigned int width, unsigned int height);
 
 
 private:
